Add TestingSupport helpers for flash and preferences
