+++ core/spec/spec_helper.rb
@@ -9,10 +9,13 @@ require 'rspec/rails'
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each {|f| require f}
 
 require 'database_cleaner'
+
 require 'spree/core/testing_support/factories'
 require 'spree/core/testing_support/env'
 require 'spree/core/testing_support/controller_requests'
 require 'spree/core/testing_support/authorization_helpers'
+require 'spree/core/testing_support/preferences'
+require 'spree/core/testing_support/flash'
 
 require 'spree/core/url_helpers'
 require 'paperclip/matchers'
@@ -55,6 +58,8 @@ RSpec.configure do |config|
   config.include FactoryGirl::Syntax::Methods
   config.include Spree::Core::UrlHelpers
   config.include Spree::Core::TestingSupport::ControllerRequests
+  config.include Spree::Core::TestingSupport::Preferences
+  config.include Spree::Core::TestingSupport::Flash
 
   config.include Paperclip::Shoulda::Matchers
 end
